add kernel_utils.c/h
