+++ ffi/examples/read-table/README.md
@@ -7,6 +7,14 @@ Simple example to show how to read and dump the data of a table using kernel's c
 
 This example is built with [cmake]. Instructions below assume you start in the directory containing this README.
 
+Note that prior to building these examples you must build `delta_kernel_ffi` (see [the FFI readme] for details). TLDR:
+```bash
+# from repo root
+$ cargo build -p delta_kernel_ffi [--release] [--features default-engine]
+# from ffi/ dir
+$ cargo build [--release] [--features default-engine]
+```
+
 ## Linux / MacOS
 
 Most likely something like this should work:
@@ -36,6 +44,11 @@ won't print out correctly.
 This example uses the `arrow-glib (c)` component from arrow to print out data. This requires
 _installing_ that component which can be non-trivial. Please see
 [here](https://arrow.apache.org/install/) to find installation instructions for your system.
+
+For macOS and homebrew this should be as easy as:
+```
+brew install apache-arrow-glib
+```
 
 If you don't want to have to install this, you can run `ccmake ..` (`cmake-gui.exe ..` on windows)
 from the `build` directory, and turn `OFF`/uncheckmark `PRINT_DATA`. Then "configure" and
